+++ src/Random/Distribution/GeometricDistribution.php
@@ -33,7 +33,7 @@ class GeometricDistribution extends AbstractDistribution
      */
     public function generate(AbstractEngine $engine)
     {
-        $result = log(1 - $engine->nextDouble()) / log(1 - $this->probability);
+        $result = log(1 - $engine->nextFloat()) / log(1 - $this->probability);
 
         return (int) $result;
     }
